Refactor the message dispatcher to a module
