+++ static/js/app-constants.js
@@ -1,0 +1,3 @@
+export const AppConstants = {
+    "CLIENT_WS_URL": window.location.host + "/ws/client"
+}
